#15 Move authorization into its own package
Authorization setup was located in the frontend while permission and role enums were located in the entity layer. Authorization was separated from the layer-architecture and moved to its own package. This looks a lot more tidy.
