+++ server/src/main/java/com/xinra/reviewcommunity/auth/Role.java
@@ -1,7 +1,8 @@
-package com.xinra.reviewcommunity.entity;
+package com.xinra.reviewcommunity.auth;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
+import com.xinra.reviewcommunity.entity.UserLevel;
 import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
