+++ server/src/main/java/com/xinra/reviewcommunity/auth/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package is dedicated authorization. Spring Security's method security is used in connection
+ * with custom enums and annotations. For authentication see
+ * {@link com.xinra.reviewcommunity.rest.conf}.
+ */
+package com.xinra.reviewcommunity.auth;
