+++ server/src/main/java/com/xinra/reviewcommunity/rest/conf/SuccessfulAuthenticationDto.java
@@ -1,7 +1,7 @@
 package com.xinra.reviewcommunity.rest.conf;
 
 import com.xinra.nucleus.service.Dto;
-import com.xinra.reviewcommunity.entity.Permission;
+import com.xinra.reviewcommunity.auth.Permission;
 import com.xinra.reviewcommunity.service.UserDto;
 import java.util.Set;
 import lombok.Data;
