+++ server/src/main/java/com/xinra/reviewcommunity/rest/conf/package-info.java
@@ -1,4 +1,5 @@
 /**
- * Contains everything related to setting up Spring for the REST API.
+ * Contains everything related to setting up Spring for the REST API. This is mainly authentication
+ * and context awareness.
  */
 package com.xinra.reviewcommunity.rest.conf;
