+++ server/src/main/java/com/xinra/reviewcommunity/service/AuthenticatedUserDto.java
@@ -1,8 +1,8 @@
 package com.xinra.reviewcommunity.service;
 
 import com.xinra.nucleus.entity.EntityPk;
-import com.xinra.reviewcommunity.entity.Permission;
-import com.xinra.reviewcommunity.entity.Role;
+import com.xinra.reviewcommunity.auth.Permission;
+import com.xinra.reviewcommunity.auth.Role;
 import java.util.Set;
 import lombok.Getter;
 import lombok.NonNull;
